minor renames
