+++ libs/mock-builder/src/lib.rs
@@ -41,9 +41,7 @@
 //! how to use this crate.
 
 /// Provide functions for register/execute calls
-//pub mod storage;
-pub mod storage_2;
-pub use storage_2 as storage;
+pub mod storage;
 
 /// Provide functions for handle fuction locations
 pub mod location;
